hit and death animation

+++ Assets/HealthPoints.cs
@@ -5,23 +5,23 @@ public struct HealthPoints
 {
     public static HealthPoints Zero = new HealthPoints(0);
 
-    [SerializeField] private ushort _value;
+    [SerializeField] private int _value;
 
-    public ushort Value { get { return _value; } }
+    public int Value { get { return _value; } }
 
-    public HealthPoints(ushort value)
+    public HealthPoints(int value)
     {
         _value = value;
     }
 
     public static HealthPoints operator +(HealthPoints hp1, HealthPoints hp2)
     {
-        return new HealthPoints((ushort)(hp1.Value + hp2.Value));
+        return new HealthPoints(hp1.Value + hp2.Value);
     }
 
     public static HealthPoints operator -(HealthPoints hp1, HealthPoints hp2)
     {
-        return new HealthPoints((ushort)(hp1.Value - hp2.Value));
+        return new HealthPoints(hp1.Value - hp2.Value);
     }
 
     public static bool operator >(HealthPoints hp1, HealthPoints hp2)
